API: Move clustering methods into separate module
